+++ Assets/Gemserk.SelectionHistory/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.8
 
 * Now scene is pinged when pressing ping for unloaded objects.
+* Now all scene references are shown with scene name prefix.
 
 # 1.0.7